+++ modules/router/src/errors.ts
@@ -14,6 +14,7 @@ export class CollateralError extends RouterError {
   static readonly reasons = {
     ChannelNotFound: "Channel not found",
     CouldNotGetOnchainDeposits: "Unable to get total deposited onchain",
+    NotInChannel: "Router is not found in channel",
     ProviderNotFound: "Provider not found",
     UnableToGetRebalanceProfile: "Could not get rebalance profile",
     TargetHigherThanThreshold: "Specified target is higher than reclaim threshold",